requirements.txt
streamlit
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-huggingface
weasyprint
markdown
python-dotenv